+++ controls.py
@@ -1,0 +1,9 @@
+from kivy.lang import Builder
+from kivy.uix.floatlayout import FloatLayout
+
+Builder.load_file("controls.kv")
+
+
+class Controls(FloatLayout):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
