+++ createGame_menu.kv
@@ -1,3 +1,6 @@
+#:import kart_simulator kart_simulator
+
+
 <CreateGame>:
     BoxLayout:
         orientation:"vertical"
@@ -35,17 +38,18 @@
                 on_text:
                     root.changeWorldSpinnerText(worlds_spinner_id.text), app.ButtonSound(), preView_ID.changePreView(root.chosen_world)
 
-            Button:
+            UpdateWorldButton:
                 size_hint: .5,.5
                 pos_hint: {'top': 1}
                 text: "Update the worlds"
-                # on_press: app.manager.push("UpdateWorlds"), app.ButtonSound()
+                on_press: self.generateUpdatedWorldsList(labelID,worlds_spinner_id)
 
             Button:
                 size_hint: .25,1
                 text: "Play"
                 background_color: (1.0, 0.0, 0.0, .75)
                 # on_press: app.instanciate_ks(root.chosen_world, root.chosen_music, root.chosen_POV), app.ButtonSound()
+                on_press: app.instanciate_ks(root.chosen_world,"Third Person")
 
         ScatterLayout:
             PreView:
@@ -57,3 +61,60 @@
         pos_hint: {"center_x":.5,"center_y":.5}
         text:""
         font_size: 32
+
+
+
+
+
+<UpdateWorlds>:
+    TextInput:
+        pos_hint: {'center_x':.5,'center_y':.5}
+        size_hint: .25, .25
+        text: ""
+        multiline: False
+        on_text_validate: root.on_text_validate(self)
+
+    ######################## A décommenter en cas d'extrême désolation dans la recherche de l'easter Egg ########################
+    Button:
+        pos_hint: {'center_x':.5,'center_y':.1}
+        size_hint: .15,.15
+        text:"EasterEgg"
+        on_press: app.manager.pop(), app.instanciate_ks(world="client/easteregg.json", music = "No music", POV="Third Person")
+    #############################################################################################################################
+    
+    Label:
+        pos_hint: {'center_x':.5,'center_y':.95}
+        bold: True
+        text: "Update the worlds"
+        font_size: 32
+        color: .8,.4,.2,1
+        size_hint: 1,.1
+        canvas.before:
+            Color:
+                rgba: 0, 0, 0, 1
+            Rectangle:
+                pos: self.pos
+                size: self.size
+
+    UpdateWorldButton:
+        pos_hint: {'center_x':.5,'center_y':.8}
+        size_hint: 1, .2
+        text: "Update the worlds now"
+        on_press:
+            self.generateUpdatedWorldsList(updateWorlds_output_id, app.manager.get_screen("MainMenu").children[0].ids.worlds_spinner_id),
+            app.ButtonSound()
+
+    Label:
+        id: updateWorlds_output_id
+        pos_hint: {'x':0,'y':0}
+        text: ""
+        font_size: 16
+        color: 1,1,1,1
+        size_hint: 1,.7
+        canvas.before:
+            Color:
+                rgba: 0,0,0,1
+            Rectangle:
+                pos: self.pos
+                size: self.size
+
