+++ layouts.py
@@ -80,192 +80,6 @@ class MyScreenManager(NavigationScreenManager):
     pass
 
 
-class EndGameMode(FloatLayout):
-    """Menu de fin de partie"""
-    pass
-
-
-class PauseMode(FloatLayout):
-    def __init__(self, width, height, music, **kwargs):
-        """Menu de pause"""
-
-        self.chosen_music = str(music)
-        self.width = width
-        self.height = height
-        super().__init__(**kwargs)
-
-    def changeMusicSpinnerText(self, text):
-        """Change le texte à afficher sur le dépliant de choix de la musique"""
-        self.chosen_music = text
-
-    def generateMusicsList(self):
-        """Génère la liste des musiques disponibles"""
-        music_list = list(music[:-4] for music in listdir("client/sounds/music"))
-        return music_list
-
-
-class KS_screen(Screen):
-    def __init__(self, world, music, POV, **kw):
-        """Screen responsable d'afficher la partie"""
-        self.musicName = self.get_musicName(music)
-        super().__init__(**kw)
-
-        self.app = App.get_running_app()
-        self.world = world
-        self.POV = POV
-        # Instantiation du canvas de jeu
-        self.game = MainWidget(self.world, POV=self.POV, parentScreen=self)
-        if self.game.theGame:
-            self.startMusic()
-            self.ids.noActionBar.add_widget(self.game)
-            # self.start_button = Button(text="start The game!", size_hint=(0.25, 0.1))
-            # self.start_button.bind(on_press=self.startingAnimation)
-            # self.ids.noActionBar.add_widget(self.start_button)
-            self.game.theGame.callOutput()
-            self.startingAnimation()
-
-    def quit(self):
-        """Nettoyage du canvas de jeu après la partie"""
-        self.game.clear()
-
-    def pauseMode(self):
-        """Appel du mode de pause"""
-        if self.musicName:
-            self.pauseMusic()
-
-        self.game.play = False
-        self.game.my_clock.unschedule(self.game.nextFrame)
-        self.pauseMenu = PauseMode(
-            width=Window.width, height=Window.height, music=self.musicName
-        )
-        self.add_widget(self.pauseMenu)
-
-    def endGameMode(self, message):
-        """Appel du mode de fin de partie"""
-        if self.musicName:
-            self.pauseMusic()
-
-        self.game.play = False
-        self.game.my_clock.unschedule(self.game.nextFrame)
-        self.endGameMenu = EndGameMode()
-        self.endGameMenu.ids.gameOverLabel_id.text = message
-        anim = (
-            Animation(font_size=74, duration=0.1)
-            + Animation(font_size=120, duration=1)
-            + Animation(font_size=120, duration=0.1)
-            + Animation(font_size=74, duration=1)
-        )
-        anim.repeat = True
-        anim.start(self.endGameMenu.ids.gameOverLabel_id)
-        self.add_widget(self.endGameMenu)
-
-    def begin_game(self, dt):
-        """Démarrage de la partie"""
-        self.game.start_theGame()
-
-    def startingAnimation(self):
-        """Création et affichage de l'animation de début de partie"""
-        # self.ids.noActionBar.remove_widget(self.start_button)
-        start_animation3 = Label(
-            text="3", font_size=0, halign="center", color=(1, 0, 1, 1)
-        )
-        start_animation2 = Label(
-            text="2", font_size=0, halign="center", color=(1, 0, 1, 1)
-        )
-        start_animation1 = Label(
-            text="1", font_size=0, halign="center", color=(1, 0, 1, 1)
-        )
-        start_animationGO = Label(
-            text="GOOOO!!!!", font_size=0, halign="center", color=(1, 0, 1, 1)
-        )
-
-        self.ids.animationLayout.add_widget(start_animation3)
-        self.ids.animationLayout.add_widget(start_animation2)
-        self.ids.animationLayout.add_widget(start_animation1)
-        self.ids.animationLayout.add_widget(start_animationGO)
-        anim = (
-            Animation(font_size=74, duration=0.5)
-            + Animation(font_size=200, duration=0.5)
-            + Animation(font_size=0, duration=0.5)
-        )
-        anim.start(start_animation3)
-        anim = (
-            Animation(duration=1.5)
-            + Animation(font_size=74, duration=0.5)
-            + Animation(font_size=200, duration=0.5)
-            + Animation(font_size=0, duration=0.5)
-        )
-        anim.start(start_animation2)
-        anim = (
-            Animation(duration=3)
-            + Animation(font_size=74, duration=0.5)
-            + Animation(font_size=200, duration=0.5)
-            + Animation(font_size=0, duration=0.5)
-        )
-        anim.start(start_animation1)
-        anim = (
-            Animation(duration=4.5)
-            + Animation(font_size=74, duration=0.5)
-            + Animation(font_size=400, duration=0.5)
-            + Animation(font_size=0, duration=0.5)
-        )
-        anim.start(start_animationGO)
-        # Appel d'une instance de l'output afin d'afficher le circuit derrière l'animation
-        Clock.schedule_once(self.begin_game, 6)
-
-    def end_game(self, endGameMessage=""):
-        """Appel du mode de fin de partie"""
-        self.endGameMode(endGameMessage)
-
-    def resumeGame(self, new_music):
-        """Reprise de la partie à la fin de la pause"""
-        self.resumeMusic()
-
-        self.game.play = True
-        self.game.my_clock.schedule_interval(self.game.nextFrame, 1 / self.game.fps)
-        self.remove_widget(self.pauseMenu)
-
-    def startMusic(self):
-        """Initialisation de la musique (avec répétitions)"""
-
-        if self.app.soundEnabled:
-            try:
-                musicPath = path.join("client/sounds/music", self.musicName) + ".wav"
-                self.music = SoundLoader.load(musicPath)
-                # self.music_pos = 0
-                self.music.volume = 1
-                self.music.play()
-                # self.music.seek(self.music_pos)
-                self.music.loop = True
-
-            except:
-                pass
-
-    def changeMusic(self, new_music):
-        """Changement de la musique"""
-        self.musicName = new_music
-
-    def pauseMusic(self):
-        """Arrêt de la musique lors du mode pause"""
-        try:
-            # self.music_pos = self.music.get_pos()     #Bug connu dans la class "sounds" de kivy, indisponible pour l'instant...
-            self.music.stop()
-            # self.music.loop = False
-        except:
-            pass
-
-    def resumeMusic(self):
-        """Reprend la musique après une pause"""
-        self.startMusic()
-
-    def get_musicName(self, music):
-        if isinstance(music, StringProperty):
-            return music.defaultvalue
-        else:
-            return music
-
-
-
 
 
 
@@ -310,94 +124,6 @@ class UpdateWorlds(FloatLayout):
             App.get_running_app().manager.push("EG1")
 
 
-class UpdateWorldButton(Button):
-    def __init__(self, **kwargs):
-        """Bouton qui met à jour dynamiquement la liste des mondes jouables"""
-        super().__init__(**kwargs)
-        self._updating = False
-
-    def generateUpdatedWorldsList(self, updateWorlds_output, worlds_spinner):
-        """Met à jour les donnés des mondes et met l'affichage à jour"""
-        if self._updating:
-            self.text = "It's already updating!"
-        else:
-            self._updating = True
-            threading.Thread(
-                target=self.generateUpdatedWorldsListTask,
-                args=(updateWorlds_output, worlds_spinner),
-            ).start()
-
-    def generateUpdatedWorldsListTask(self, updateWorlds_output, worlds_spinner):
-        """Récupère les données des mondes et met à jour l'affichage"""
-        updateWorlds_output.text = "\nUpdating the worlds ...\n"
-        worldsInfo = {}
-        session = requests.Session()
-        try:
-            worlds = session.get(
-                "https://lj44.ch/creator/kart/worldsjson",
-                params={"id": True, "version": True, "name": True},
-                timeout=1,
-            )
-        except requests.ConnectionError:
-            updateWorlds_output.text += "ERROR: The server in unreachable"
-            self._updating = False
-            self.text = "Update the worlds now"
-        else:
-            for world in worlds.json():
-                worldsInfo[world["name"]] = {"id": world["id"], "version": world["version"]}
-            try:
-                f = open("client/worlds.json", "r")
-            except FileNotFoundError:
-                f = open("client/worlds.json", "w")
-                f.write("{}")
-                f.close()
-                f = open("client/worlds.json", "r")
-
-            try:
-                savedWorld = json.load(f)
-                # mise à jour des mondes déjà téléchargés
-                for name, data in savedWorld.items():
-                    if name in worldsInfo:
-                        if data["version"] != worldsInfo[name]["version"]:
-                            updateWorlds_output.text += f"Updating world {name} ... "
-                            with open(f"client/worlds/{name}.json", "w") as worldJSON:
-                                worldJSON.write(
-                                    session.get(
-                                        f"https://lj44.ch/creator/kart/worlds/{worldsInfo[name]['id']}/fabric"
-                                    ).text
-                                )
-                            updateWorlds_output.text += "done!\n"
-                            data["version"] = worldsInfo[name]["version"]
-                    else:
-                        updateWorlds_output.text += f"Deleting world {name} ... "
-                        os.remove(f"client/worlds/{name}.json")
-                        updateWorlds_output.text += "done!\n"
-                # téléchargement des autres
-                for name, data in worldsInfo.items():
-                    if name not in savedWorld:
-                        updateWorlds_output.text += f"Downloading world {name} ... "
-                        with open(f"client/worlds/{name}.json", "w") as worldJSON:
-                            worldJSON.write(
-                                session.get(
-                                    f"https://lj44.ch/creator/kart/worlds/{worldsInfo[name]['id']}/fabric"
-                                ).text
-                            )
-                        updateWorlds_output.text += "done!\n"
-            finally:
-                f.close()
-
-            with open("client/worlds.json", "w") as f:
-                json.dump(worldsInfo, f)
-
-            worlds_spinner.values = [world[:-5] for world in listdir("client/worlds")]
-            updateWorlds_output.text += "All worlds are up to date!"
-            self._updating = False
-            self.text = "Update the worlds now"
-            if App.get_running_app().soundEnabled:
-                sound = SoundLoader.load("client/sounds/success-sound-effect.mp3")
-                sound.volume = 0.5
-                sound.play()
-
 
 ##########################################################################
 
@@ -426,9 +152,7 @@ class PasswordScreen(FloatLayout):
             self.app.instanciate_ks(world="client/easteregg.json", music="No music", POV="Thrid Person")
 
 
-class Controls(FloatLayout):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
+
