+++ menu.kv
@@ -13,167 +13,8 @@
 #:import kart_settings kart_settings
 #:import settings settings
 #:import music_settings music_settings
+#:import controls controls
 # Création de l'interface de tous les screens
-
-
-# <MainMenu2>:
-#     GridLayout:
-#     # Image de fond
-#         canvas.before:
-#             BorderImage:
-#                 border: 10, 10, 10, 10
-#                 source: 'client/Images/background.jpg'
-#                 pos: self.pos
-#                 size: self.size
-#     Label:
-#         bold: True
-#         text: "Kart Simulator"
-#         font_size: 50
-#         color: 1,0,0,1
-#         pos_hint: {'x':0,'y':.85}
-#         size_hint: 1,.15
-#         canvas.before:
-#             Color:
-#                 rgba: 0, 0, 0, .75
-#             Rectangle:
-#                 pos: self.pos
-#                 size: self.size
-#     Button:
-#         pos_hint: {'x':.3,'y':.3}
-#         size_hint: .4, .3
-#         text: "Play"
-#         background_color: (1.0, 0.0, 0.0, .75)
-#         on_press: app.instanciate_ks(root.chosen_world, root.chosen_music, root.chosen_POV), app.ButtonSound()
-
-#     Button:
-#         pos_hint: {'x':.8,'y':.8}
-#         size_hint: .2, .05
-#         text: "Sign in / Log in"
-#         on_press: app.manager.push("connection"), app.ButtonSound()
-
-#     Button:
-#         pos_hint: {'x':.8,'y':.75}
-#         size_hint: .2, .05
-#         text: "Create/edit a world"
-#         on_press: webbrowser.open("https://lj44.ch/creator/kart"), app.ButtonSound()
-
-#     Button:
-#         pos_hint: {'x':.8,'y':.7}
-#         size_hint: .2, .05
-#         text: "Multiplayer"
-#         on_press: app.manager.push("Multiplayer"), app.ButtonSound()
-
-
-#     Button:
-#         pos_hint: {'x':0,'y':.7}
-#         size_hint: .2, .05
-#         text: "Update the worlds"
-#         on_press: app.manager.push("UpdateWorlds"), app.ButtonSound()
-
-
-#     Spinner:
-#         id: worlds_spinner_id
-#         pos_hint: {'x':0,'y':.6}
-#         size_hint: .2, .1
-#         text: "Chose your world !"
-#         values: root.generateWorldsList()
-#         on_text:
-#             root.changeWorldSpinnerText(worlds_spinner_id.text), app.ButtonSound(), preView_ID.changePreView(root.chosen_world)
-
-#     Spinner:
-#         id: music_spinner_id
-#         pos_hint: {'x':.8,'y':.6}
-#         size_hint: .2, .1
-#         text: "Choose your music !"
-#         values: root.generateMusicsList()
-#         on_text:
-#             root.changeMusicSpinnerText(music_spinner_id.text), app.ButtonSound()
-
-#     Spinner:
-#         id: POV_spinner_id
-#         pos_hint: {'x':.8,'y':.3}
-#         size_hint: .2, .1
-#         text: "Choose your POV !"
-#         values: ["First Person","Third Person"]
-#         on_text:
-#             root.changePOVSpinnerText(POV_spinner_id.text), app.ButtonSound()
-
-#     Button:
-#         id: prewViewButton_ID
-#         pos_hint: {'x':.2,'y':.6}
-#         size_hint: .08, .1
-#         text: "Preview"
-#         on_press: preView_ID.updatePreviewMode(), preView_ID.changePreView(root.chosen_world), app.ButtonSound()
-
-#     ScatterLayout:
-#         pos: 0,0
-#         size_hint: .2, .2
-#         PreView:
-#             maxSize:(100,100)
-#             id: preView_ID
-
-
-<KS_screen>:
-    name: "Kart_Simulator"
-    layout_id: noActionBar
-    size: app.windowSize()
-
-    ScatterLayout:
-        id: noActionBar
-        pos: 0,0
-        size: root.size
-        do_translation: False
-        do_rotation: False
-        do_scale: False
-
-    FloatLayout:
-        id: animationLayout
-        pos: 0,0
-        size: app.windowSize()
-        # canvas.before:
-        #     Color: 
-        #         rgba: 0,0,0,1
-        #     Rectangle:
-        #         size: self.size[0]/2,self.size[1]/2
-
-    
-    GridLayout:
-        cols: 2
-        rows: 3
-        pos_hint: {'x': .45, 'y': .9}
-        size_hint: .15, .1
-        canvas.before:
-            Color:
-                rgba: 0,0,0,.5
-            Rectangle:
-                pos: self.pos
-                size: self.size
-
-        Label:
-            text: "Laps"
-            color: 0.4, 1, 0.4, .5
-
-        Label:
-            id: laps_id
-            text: "0/1"
-            color: 0.4, 1, 0.4, .5
-
-        Label:
-            text: "Gates"
-            color: 0.4, 1, 0.4, .5
-
-        Label:
-            id: gates_id
-            text: "0/2"
-            color: 0.4, 1, 0.4, .5
-        Label:
-            text: "Timer"
-            color: 0.4, 1, 0.4, .5
-        Label:
-            id: timer_id
-            text: "00:00:00"
-            color: 0.4, 1, 0.4, .5
-
 
 
 <MyScreenManager>:
@@ -278,15 +119,4 @@
         BoxLayoutWithActionBar:
             title: "Credits"
             Credits:
-
-
-
-
-
-
-<MenuTabbedPanel@TabbedPanel>:
-    do_default_tab: False
-    TabbedPanelItem:
-        text: "Log in"
-        GameConnection:
