+++ music_settings.py
@@ -42,7 +42,7 @@ class MusicSettings(GridLayout):
     def updateUserSettings(self, instance):
         """Met à jour les paramètres du comte du joueur."""
         app = App.get_running_app()
-        data = app.session.get("http://localhost:5000/auth/myaccount/kart.json")
+        data = app.session.get("http://localhost:5000/auth/myaccount/kart.json").json()
         data["music"] = self._chosenMusic
         app.session.put("http://localhost:5000/auth/myaccount/kart.json", data)
